+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Reuse `<SearchResults>` component from `stripes-data-transfer-components` rep (UIDATIMP-581)
 * Cover <BooleanActionField> component with tests (UIDATIMP-700)
 * Reuse `<SearchForm>` component from `stripes-data-transfer-components` rep (UIDATIMP-582)
+* Cover `<ActionMenu>` component with tests (UIDATIMP-699)
 
 ### Bugs fixed:
 * Fix Accessibility problems for settings/data-import/match-profiles (lists must only directly contain li elements) (UIDATIMP-452)